fix test, agents, utils for v3 and v4 use same files
