+++ sol080/contracts/oceanv4/test/test_swap_fees.py
@@ -1,7 +1,6 @@
 import brownie
 
-import sol080.contracts.oceanv4.oceanv4util
-from sol080.contracts.oceanv4.oceanv4util import ROUTER 
+from sol080.contracts.oceanv4 import oceanv4util
 from util.base18 import toBase18
 from util.constants import BROWNIE_PROJECT080, OPF_ACCOUNT, GOD_ACCOUNT
 from util.globaltokens import OCEANtoken
@@ -68,15 +67,17 @@ def test_exactAmountIn_fee():
 
 def _deployBPool():
     brownie.chain.reset()
-    router = ROUTER()
-    fundOCEANFromAbove(address0, toBase18(10000.0))
-    dataNFT = sol080.contracts.oceanv4.oceanv4util.createDataNFT(
+    router = oceanv4util.deployRouter(account0)
+    fundOCEANFromAbove(address0, toBase18(10000))
+    createdDataNFT = oceanv4util.createDataNFT(
         "dataNFT", "DATANFTSYMBOL", account0, router
     )
-    datatoken = sol080.contracts.oceanv4.oceanv4util.create_datatoken_from_dataNFT(
+    dataNFT = createdDataNFT[0]
+    datatoken = oceanv4util.createDatatokenFromDataNFT(
         "DT", "DTSYMBOL", 10000, dataNFT, account0
     )
-    pool = sol080.contracts.oceanv4.oceanv4util.create_BPool_from_datatoken(
-        datatoken, 1000, 2000, account0
+    erc721_factory = createdDataNFT[1]
+    pool = oceanv4util.createBPoolFromDatatoken(
+        datatoken, 1000, 2000, account0, erc721_factory
     )
     return pool
