+++ sol080/contracts/oceanv4/test/test_vesting.py
@@ -1,10 +1,6 @@
 import brownie
 
-import sol080.contracts.oceanv4.oceanv4util
-from sol080.contracts.oceanv4.oceanv4util import (
-    ROUTER,
-    SIDESTAKING,
-)
+from sol080.contracts.oceanv4 import oceanv4util
 from util.base18 import toBase18
 from util.constants import BROWNIE_PROJECT080, OPF_ACCOUNT, GOD_ACCOUNT, ZERO_ADDRESS
 from util.globaltokens import fundOCEANFromAbove
@@ -37,7 +33,8 @@ def test_fail_to_mint():
 def test_vesting_amount():
     pool = _deployBPool()
     datatoken = BROWNIE_PROJECT080.ERC20Template.at(pool.getDataTokenAddress())
-    sideStaking = SIDESTAKING()
+    sideStakingAddress = pool.getController()
+    sideStaking = BROWNIE_PROJECT080.SideStaking.at(sideStakingAddress)
     assert sideStaking.getvestingAmount(datatoken.address) == toBase18(1000)
 
     # TODO: https://github.com/oceanprotocol/contracts/blob/v4main_postaudit/test/flow/Vesting.test.js#L292
@@ -46,15 +43,17 @@ def test_vesting_amount():
 
 def _deployBPool():
     brownie.chain.reset()
-    router = ROUTER()
-    fundOCEANFromAbove(address0, toBase18(100000))
-    dataNFT = sol080.contracts.oceanv4.oceanv4util.createDataNFT(
+    router = oceanv4util.deployRouter(account0)
+    fundOCEANFromAbove(address0, toBase18(10000))
+    createdDataNFT = oceanv4util.createDataNFT(
         "dataNFT", "DATANFTSYMBOL", account0, router
     )
-    datatoken = sol080.contracts.oceanv4.oceanv4util.create_datatoken_from_dataNFT(
+    dataNFT = createdDataNFT[0]
+    datatoken = oceanv4util.createDatatokenFromDataNFT(
         "DT", "DTSYMBOL", 10000, dataNFT, account0
     )
-    pool = sol080.contracts.oceanv4.oceanv4util.create_BPool_from_datatoken(
-        datatoken, 1000, 2000, account0
+    erc721_factory = createdDataNFT[1]
+    pool = oceanv4util.createBPoolFromDatatoken(
+        datatoken, 1000, 2000, account0, erc721_factory
     )
     return pool
